pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wabotii"]

[project]
name = "wabotii"
version = "0.1.0"
description = "WhatsApp video downloader bot using WAHA and FastAPI"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Don Folayan", email = "donfolayan@yahoo.com"}
]
license = {text = "MIT"}
keywords = ["whatsapp", "bot", "video", "downloader", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications",
    "Topic :: Internet",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic-settings>=2.1.0",
    "pydantic>=2.5.3",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.2",
    "requests>=2.31.0",
    "yt-dlp>=2024.1.0",
    "cloudinary>=1.36.0",
    "gunicorn>=21.2.0",
    "sqlalchemy>=2.0.23",
    "structlog>=24.1.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.2",
    "ruff>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "black==23.12.1",
    "mypy==1.8.0",
    "pre-commit==3.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
